Introduce new method find_restaurant
The method find_restaurants is used to return a list of all the
restaurants from the db, while the new method find_restaurant is used to
return a single result.

The new method is using find_one instead of find. Together with the
removal of the if statement, this will make the retrieval of the
result(s) a bit faster.