Add install stage in gitlab-ci and a job to install the python deps
The install stage is used for installing the API app's dependencies,
which are afterwards cached.

The docker image that is used for building is a custom made, that is
based on the python:3.7 official image, including the virtualenv
package.

In a later commit, the dependencies will be used by the API app's
Dockerfile. This way, the dependencies will not be reinstalled in the
CI/CD on every docker build, reducing the build time.