Add additional Dockerfile for the API and build options in docker-compose
In case that the user does not have access to the private registry, I am
adding a commented out build section for both images.

Additionally, I am adding another Dockerfile for the API, which installs
the requirements during the docker build.

With the above two, a user can run `docker-compose up` and have a
development environment, without the need of doing anything else on the
CI or locally.