Add kubernetes files
There is a new namespace called "restaurants":

The API app consists of a secrets file (which contains only the
MONGO_URI env var), a deployment and a service with NodePort, to be able
to bind the port 8080 externally.

The mongodb consists of a secrets file (with the root credentials, and
the api app db user's credentials), a configmap with the db name, a
deployment and a service with ClusterIP, as only the API needs to
contact it.

The images used are public ones, as I was having a hard time to be able
to register "kind" to the private gitlab registry.