Add SOLUTIONS.md and improve .gitignore
This firt commit adds an explanatory file SOLUTIONS.md where I'll be
explaining  the steps I take in order to solve the challenges, and
also deletes some dupes in .gitignore. I think it is way better to have
a customized .gitignore, even though it was first generated using a
templating system.